backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.10.1

# File Upload and Processing
python-multipart==0.0.6
PyMuPDF==1.23.8

# AI and ML
openai==1.3.7
qdrant-client==1.7.0
scikit-learn==1.3.2

# HTTP and Web
requests==2.31.0
httpx==0.25.2

# Web Search APIs
google-search-results==2.4.2  # SerpAPI

# Caching and Performance
redis==5.0.1
aioredis==2.0.1

# Monitoring and Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Additional Utilities
aiofiles==23.2.0
python-slugify==8.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
